+++ src/scripts/systems/cosmere-rpg.js
@@ -1,0 +1,34 @@
+import CONSTANTS from '../constants.js';
+import { processNotes } from "../textUtil.js";
+
+export async function cosmereRPG(actor, displayImages) {
+    // Using a guard here looks cleaner
+    if (!actor) {
+        return null;
+    }
+
+    const actorIsOwner = actor.isOwner ?? true;
+
+    switch (actor.type) {
+        case "character":
+            if (game.settings.get(CONSTANTS.MODULE_ID, 'displayPC')) {
+                return await getCharacterNotes(displayImages, actor, actorIsOwner);
+            } else {
+                return null;
+            }
+        case "adversary":
+            if (game.settings.get(CONSTANTS.MODULE_ID, 'displayNPC')) {
+                return await getCharacterNotes(displayImages, actor, actorIsOwner);
+            } else {
+                return null;
+            }
+        default:
+            return null;
+    }
+}
+
+async function getCharacterNotes(displayImages, actor, actorIsOwner) {
+    const notes = actor.system?.notes;
+
+    return await processNotes(notes, actorIsOwner, displayImages);
+}