+++ src/scripts/systems/d35e.js
@@ -1,0 +1,57 @@
+import CONSTANTS from '../constants.js';
+import { processNotes } from "../textUtil.js";
+
+export async function d35e(actor, displayImages) {
+    // Using a guard here looks cleaner
+    if (!actor) {
+        return null;
+    }
+
+    const actorIsOwner = actor.isOwner ?? true;
+
+    switch (actor.type) {
+        case "character":
+            if (game.settings.get(CONSTANTS.MODULE_ID, 'displayPC')) {
+                return await getCharacterNotes(displayImages, actor, actorIsOwner);
+            } else {
+                return null;
+            }
+        case "npc":
+            if (game.settings.get(CONSTANTS.MODULE_ID, 'displayNPC')) {
+                return await getNpcNotes(displayImages, actor, actorIsOwner);
+            } else {
+                return null;
+            }
+        case "object":
+            if (game.settings.get(CONSTANTS.MODULE_ID, 'displayNPC')) {
+                return await getNpcNotes(displayImages, actor, actorIsOwner);
+            } else {
+                return null;
+            }
+        case "trap":
+            if (game.settings.get(CONSTANTS.MODULE_ID, 'displayNPC')) {
+                return await getNpcNotes(displayImages, actor, actorIsOwner);
+            } else {
+                return null;
+            }
+        default:
+            return null;
+    }
+}
+
+async function getCharacterNotes(displayImages, actor, actorIsOwner) {
+    return await processNotes(actor.system?.details?.notes.value, actorIsOwner, displayImages);
+}
+
+async function getNpcNotes(displayImages, actor, actorIsOwner) {
+    const publicNotes = "";
+    const privateNotes = actor.system?.details?.notes.value;
+    let notes = publicNotes;
+
+    if (!game.settings.get(CONSTANTS.MODULE_ID, 'hidePrivateNotes') && game.user.isGM && privateNotes) {
+        notes += "<div class=\"token-note-hover-hud-h3\">GM Notes</div>";
+        notes += privateNotes;
+    }
+
+    return await processNotes(notes, actorIsOwner, displayImages);
+}