+++ src/scripts/systems/dnd5e.js
@@ -26,8 +26,22 @@ export async function dnd5e(actor, displayImages) {
         return null;
       }
     case "group":
-      return await processNotes(actor.system?.description?.full, actorIsOwner, displayImages);
+      return await getGroupNotes(displayImages, actor, actorIsOwner);
     default:
       return null;
   }
+}
+
+async function getGroupNotes(displayImages, actor, actorIsOwner) {
+  let groupNoteArray = "<div class=\"token-note-hover-hud-h3\">Group Members</div>";
+  for (let i = 0; i < actor.system.members.length; i += 1) {
+      groupNoteArray = groupNoteArray.concat("<p>");
+      groupNoteArray = groupNoteArray.concat(actor.system.members[i].actor?.name);
+      groupNoteArray = groupNoteArray.concat("</p>");
+  }
+  let publicNotes = "<div class=\"token-note-hover-hud-h3\">Group Details</div>";
+  publicNotes = publicNotes.concat(actor.system?.description?.summary, actor.system?.description?.full);
+  let notes = groupNoteArray.concat(publicNotes);
+
+  return await processNotes(notes, actorIsOwner, displayImages);
 }