+++ src/scripts/systems/fallout.js
@@ -1,7 +1,7 @@
 import CONSTANTS from '../constants.js';
 import { processNotes } from "../textUtil.js";
 
-export async function drawsteel(actor, displayImages) {
+export async function fallout(actor, displayImages) {
     // Using a guard here looks cleaner
     if (!actor) {
         return null;
@@ -10,7 +10,7 @@ export async function drawsteel(actor, displayImages) {
     const actorIsOwner = actor.isOwner ?? true;
 
     switch (actor.type) {
-        case "hero":
+        case "character":
             if (game.settings.get(CONSTANTS.MODULE_ID, 'displayPC')) {
                 return await getCharacterNotes(displayImages, actor, actorIsOwner);
             } else {
@@ -22,19 +22,31 @@ export async function drawsteel(actor, displayImages) {
             } else {
                 return null;
             }
-        case "object":
+        case "creature":
             if (game.settings.get(CONSTANTS.MODULE_ID, 'displayNPC')) {
                 return await getCharacterNotes(displayImages, actor, actorIsOwner);
             } else {
                 return null;
             }
-        case "party":
+        case "robot":
             if (game.settings.get(CONSTANTS.MODULE_ID, 'displayPC')) {
-                return await getPartyNotes(displayImages, actor, actorIsOwner);
+                return await getCharacterNotes(displayImages, actor, actorIsOwner);
             } else {
                 return null;
             }
-        case "retainer":
+        case "scavenging_location":
+            if (game.settings.get(CONSTANTS.MODULE_ID, 'displayNPC')) {
+                return await getLocationNotes(displayImages, actor, actorIsOwner);
+            } else {
+                return null;
+            }
+        case "settlement":
+            if (game.settings.get(CONSTANTS.MODULE_ID, 'displayNPC')) {
+                return await getCharacterNotes(displayImages, actor, actorIsOwner);
+            } else {
+                return null;
+            }
+        case "vehicle":
             if (game.settings.get(CONSTANTS.MODULE_ID, 'displayNPC')) {
                 return await getCharacterNotes(displayImages, actor, actorIsOwner);
             } else {
@@ -46,28 +58,13 @@ export async function drawsteel(actor, displayImages) {
 }
 
 async function getCharacterNotes(displayImages, actor, actorIsOwner) {
-    const publicNotes = actor.system?.biography?.value;
-    const privateNotes = actor.system?.biography?.director;
-    let notes = publicNotes;
-
-    if (!game.settings.get(CONSTANTS.MODULE_ID, 'hidePrivateNotes') && game.user.isGM && privateNotes) {
-        notes += "<div class=\"token-note-hover-hud-h3\">Director Notes</div>";
-        notes += privateNotes;
-    }
+    const notes = actor.system?.biography;
 
     return await processNotes(notes, actorIsOwner, displayImages);
 }
 
-async function getPartyNotes(displayImages, actor, actorIsOwner) {
-    let partyNoteArray = "<div class=\"token-note-hover-hud-h3\">Party Members</div>";
-    for (let i = 0; i < actor.system.members.contents.length; i += 1) {
-        partyNoteArray = partyNoteArray.concat("<p>");
-        partyNoteArray = partyNoteArray.concat(actor.system.members.contents[i].actor.name);
-        partyNoteArray = partyNoteArray.concat("</p>");
-    }
-    let publicNotes = "<div class=\"token-note-hover-hud-h3\">Party Details</div>";
-    publicNotes = publicNotes.concat(actor.system?.description?.value);
-    let notes = partyNoteArray.concat(publicNotes);
+async function getLocationNotes(displayImages, actor, actorIsOwner) {
+    const notes = actor.system?.notes;
 
     return await processNotes(notes, actorIsOwner, displayImages);
 }