+++ src/scripts/systems/starwarsffg.js
@@ -1,0 +1,60 @@
+import CONSTANTS from '../constants.js';
+import { processNotes } from "../textUtil.js";
+
+export async function starwarsffg(actor, displayImages) {
+    // Using a guard here looks cleaner
+    if (!actor) {
+        return null;
+    }
+
+    const actorIsOwner = actor.isOwner ?? true;
+
+    switch (actor.type) {
+        case "character":
+            if (game.settings.get(CONSTANTS.MODULE_ID, 'displayPC')) {
+                return await getCharacterNotes(displayImages, actor, actorIsOwner);
+            } else {
+                return null;
+            }
+        case "vehicle":
+            if (game.settings.get(CONSTANTS.MODULE_ID, 'displayPC')) {
+                return await getCharacterNotes(displayImages, actor, actorIsOwner);
+            } else {
+                return null;
+            }
+        case "homestead":
+            if (game.settings.get(CONSTANTS.MODULE_ID, 'displayPC')) {
+                return await getCharacterNotes(displayImages, actor, actorIsOwner);
+            } else {
+                return null;
+            }
+        case "minion":
+            if (game.settings.get(CONSTANTS.MODULE_ID, 'displayNPC')) {
+                return await getNpcNotes(displayImages, actor, actorIsOwner);
+            } else {
+                return null;
+            }
+        case "nemesis":
+            if (game.settings.get(CONSTANTS.MODULE_ID, 'displayNPC')) {
+                return await getNpcNotes(displayImages, actor, actorIsOwner);
+            } else {
+                return null;
+            }
+        case "rival":
+            if (game.settings.get(CONSTANTS.MODULE_ID, 'displayNPC')) {
+                return await getNpcNotes(displayImages, actor, actorIsOwner);
+            } else {
+                return null;
+            }
+        default:
+            return null;
+    }
+}
+
+async function getCharacterNotes(displayImages, actor, actorIsOwner) {
+    return await processNotes(actor.system?.biography, actorIsOwner, displayImages);
+}
+
+async function getNpcNotes(displayImages, actor, actorIsOwner) {
+    return await processNotes(actor.system?.biography, actorIsOwner, displayImages);
+}